+++ sources/packages/web/src/services/http/dto/Application.dto.ts
@@ -1,0 +1,23 @@
+import {
+  ApplicationExceptionStatus,
+  ApplicationStatus,
+  OfferingStatus,
+  ProgramInfoStatus,
+  SuccessWaitingStatus,
+} from "@/types";
+
+export interface InProgressApplicationDetailsAPIOutDTO {
+  id: number;
+  applicationStatus: ApplicationStatus;
+  pirStatus: ProgramInfoStatus;
+  PIRDeniedReason?: string;
+  offeringStatus?: OfferingStatus;
+  exceptionStatus?: ApplicationExceptionStatus;
+  parent1IncomeVerificationStatus?: SuccessWaitingStatus;
+  parent2IncomeVerificationStatus?: SuccessWaitingStatus;
+  partnerIncomeVerificationStatus?: SuccessWaitingStatus;
+  studentIncomeVerificationStatus?: SuccessWaitingStatus;
+  parent1Info?: SuccessWaitingStatus;
+  parent2Info?: SuccessWaitingStatus;
+  partnerInfo?: SuccessWaitingStatus;
+}
