+++ sources/packages/web/src/services/http/dto/index.ts
@@ -16,3 +16,4 @@ export * from "@/services/http/dto/ProgramInfoRequest.dto";
 export * from "@/services/http/dto/StudentAccountApplication.dto";
 export * from "@/services/http/dto/Note.dto";
 export * from "@/services/http/dto/Restriction.dto";
+export * from "@/services/http/dto/Application.dto";
