+++ sources/packages/web/src/types/contracts/student/ApplicationContract.ts
@@ -3,8 +3,6 @@ import {
   COEStatus,
   AssessmentStatus,
   OfferingIntensity,
-  OfferingStatus,
-  ApplicationExceptionStatus,
 } from "@/types";
 
 export interface SaveStudentApplicationDto {
@@ -248,20 +246,4 @@ export interface ApplicationDetailHeader {
 export enum SuccessWaitingStatus {
   Success = "Success",
   Waiting = "Waiting",
-}
-
-export interface InProgressApplicationDetailsAPIOutDTO {
-  id: number;
-  applicationStatus: ApplicationStatus;
-  pirStatus: ProgramInfoStatus;
-  PIRDeniedReason?: string;
-  offeringStatus?: OfferingStatus;
-  exceptionStatus?: ApplicationExceptionStatus;
-  parent1IncomeVerificationStatus?: SuccessWaitingStatus;
-  parent2IncomeVerificationStatus?: SuccessWaitingStatus;
-  partnerIncomeVerificationStatus?: SuccessWaitingStatus;
-  studentIncomeVerificationStatus?: SuccessWaitingStatus;
-  parent1Info?: SuccessWaitingStatus;
-  parent2Info?: SuccessWaitingStatus;
-  partnerInfo?: SuccessWaitingStatus;
 }
